==> pipeline/04_call_broad_cn.ipynb <==
import pandas as pd 
import numpy as np 

df_fish = '/home/jiageng/Documents/fhr/data/SeqFISH Files_MMRF_CoMMpass_IA22_genome_gatk_cna_seqFISH.tsv'
df_fish = pd.read_csv(df_fish, sep='\t').assign(PUBLIC_ID=lambda x: x['SAMPLE'].str.extract(r'(MMRF_\d{4})')).sort_values('SAMPLE').groupby('PUBLIC_ID').head(1)
print('df_fish',df_fish.shape)


matfish = df_fish.filter(regex = r'PUBLIC_ID|SeqWGS_Cp_\d+[pq]\d+$').set_index('PUBLIC_ID')
print(matfish.shape)

matfish.to_csv('../matrices/broad_cn_matrix.tsv',sep='\t')
