==> pipeline/04_feature_selection_gene_expr.ipynb <==
# Feature selection

import pandas as pd
import numpy as np
import torch
import os
os.chdir('/home/jiageng/Documents/fhr/pipeline/')
import snf
from sklearn.feature_selection import *

fhr_data = pd.read_csv('../annotations/fhr-annotations.tsv',sep='\t').set_index('PUBLIC_ID').query('risk != -1')
fhr_data['risk'] += 1

subtype = 'mirna' # gene (proteincoding), mirna, lncrna

# subset to samples with fhr labels
if subtype == 'gene':
    # protein coding genes
    rnamat = pd.read_csv('../matrices/gene_exp_matrix.tsv',sep='\t').set_index('PUBLIC_ID').drop(columns=['SAMPLE'])
    ensgid = pd.read_csv('../annotations/ensgid-autosomal-proteincoding.txt',sep='\t').set_index('Gene stable ID')
elif subtype == 'mirna':
    # miRNA
    rnamat = pd.read_csv('../matrices/mirna_exp_matrix.tsv',sep='\t').set_index('PUBLIC_ID').drop(columns=['SAMPLE'])
    ensgid = pd.read_csv('/home/jiageng/Documents/fhr/annotations/ensgid-autosomal-mirna.txt', sep='\t').set_index('Gene stable ID')
elif subtype == 'lncrna':
    # lncRNA
    rnamat = pd.read_csv('../matrices/lncrna_exp_matrix.tsv',sep='\t').set_index('PUBLIC_ID').drop(columns=['SAMPLE'])
    ensgid = pd.read_csv('/home/jiageng/Documents/fhr/annotations/ensgid-autosomal-lncrna.txt', sep='\t').set_index('Gene stable ID')
else:
    raise('Unknown subtype:', subtype)

print(subtype, rnamat.shape)

public_ids = list(set(fhr_data.index).intersection(set(rnamat.index)))
print(len(public_ids))

def stdNormalize(df):
    numeric_df = df.select_dtypes(include=[np.number])
    std = numeric_df.std().fillna(1)
    mean = numeric_df - numeric_df.mean()
    numeric_df_norm = mean / std
    nonnumeric_df = df.select_dtypes(exclude=[np.number])
    df_norm = pd.concat([nonnumeric_df, numeric_df_norm], axis=1)
    return df_norm

X = stdNormalize(rnamat).loc[public_ids]
y = fhr_data.loc[public_ids,'risk']

Feature selection by FWE significance of 0.05

miRNA gives only 3 genes. Too stringent

fwe = SelectFwe(f_classif, alpha=0.05).fit(X, y)
print(len(fwe.get_feature_names_out()))
ensgid.loc[fwe.get_feature_names_out()]

rnamat[fwe.get_feature_names_out()].to_csv(f'../matrices/{subtype}_exp_matrix_fwe.tsv',sep='\t')

Feature selection by specifying FDR of 0.05

fdr = SelectFdr(f_classif, alpha=0.05).fit(X, y)
print(len(fdr.get_feature_names_out()))
ensgid.loc[fdr.get_feature_names_out()]

rnamat[fdr.get_feature_names_out()].to_csv(f'../matrices/{subtype}_exp_matrix_fdr.tsv',sep='\t')

Feature selection by specifying FPR of 0.05

fpr = SelectFpr(f_classif, alpha=0.05).fit(X, y)
print(len(fpr.get_feature_names_out()))
ensgid.loc[fpr.get_feature_names_out()]

rnamat[fpr.get_feature_names_out()].to_csv(f'../matrices/{subtype}_exp_matrix_fpr.tsv',sep='\t')

Feature selection by K best method

k=20
kbest = SelectKBest(f_classif, k=k).fit(X,y)
rnamat[kbest.get_feature_names_out()].to_csv(f'../matrices/{subtype}_exp_matrix_k{k}.tsv',sep='\t')
