==> pipeline/07_gcn_gene_exp.ipynb <==
import pandas as pd
import numpy as np
import torch
import os
# os.chdir('/home/jiageng/Documents/fhr/pipeline/')
os.chdir('/home/jiageng/Documents/fhr/pygcn/pygcn')
import snf

def stdNormalize(df):
    std = df.std().fillna(1)
    mean = df - df.mean()
    df_norm = mean / std
    return np.array(df_norm)

def rowNormalize(mx):
    """Row-normalize matrix"""
    rowsum = np.array(mx.sum(1))
    r_inv = np.power(rowsum, -1).flatten()
    r_inv[np.isinf(r_inv)] = 0.
    r_mat_inv = np.diag(r_inv)
    mx = r_mat_inv.dot(mx)
    return mx

from scipy.sparse import coo_matrix

def sparse_tensor_from_arr(arr):
    """Convert numpy array to torch sparse tensor"""
    # Convert numpy array to scipy sparse matrix
    sparse_sp = coo_matrix(arr)
    
    # Convert scipy sparse matrix to torch sparse tensor
    sparse_tensor = torch.sparse_coo_tensor(
        torch.LongTensor([sparse_sp.row, sparse_sp.col]),
        torch.FloatTensor(sparse_sp.data),
        torch.Size(sparse_sp.shape)
    )
    
    return sparse_tensor

Prepare labels

df_labels = pd.read_csv('/home/jiageng/Documents/fhr/annotations/fhr-annotations.tsv',sep='\t').set_index('PUBLIC_ID').query('risk != -1')
df_labels['risk'] += 1

Prepare features

# subset to samples with fhr labels
data = pd.read_csv('/home/jiageng/Documents/fhr/matrices/gene_exp_matrix_k500.tsv',sep='\t').set_index('PUBLIC_ID')
if 'SAMPLE' in data.columns:
    data = data.drop(columns=['SAMPLE'])
print(data.shape)

Use all samples with features, handle the missing labels

# use all rna seq samples
public_ids = data.index
print(len(public_ids))

feature_norm_mtd = '' 

if feature_norm_mtd == 'stdnorm':
    features = torch.tensor(stdNormalize(data.loc[public_ids]),dtype=torch.float32)
elif feature_norm_mtd == 'rownorm':
    features = torch.tensor(rowNormalize(data.loc[public_ids].values),dtype=torch.float32)
else:
    # default is to use raw log tpm+1 values
    features = torch.tensor(data.loc[public_ids].values,dtype=torch.float32)
print(features.size())

risk = df_labels.reindex(pd.Index(public_ids)).loc[public_ids]['risk'].fillna(-1).astype(int)
labels = torch.tensor(risk.values,dtype=torch.long)

# missing values are set to -1
risk.value_counts()

idx_labeled = np.where(risk != -1)[0]
idx_unlabeled = np.where(risk == -1)[0]
idx_val = torch.tensor(idx_labeled[::3])
idx_test = idx_val
idx_train = torch.tensor([idx for idx in idx_labeled if idx not in idx_val])
print('Un-labeled',len(idx_unlabeled))
print('Labeled train',len(idx_train))
print('Labeled val',len(idx_val))

# 1 = SR, 2 = GHR, 3 = FHR
risk.iloc[idx_train].value_counts()

# 1 = SR, 2 = GHR, 3 = FHR
# idx_val and idx_test are the same for now
risk.iloc[idx_val].value_counts()

Prepare adjacency matrix

import snf
aff = snf.make_affinity(stdNormalize(data), metric='cosine', K=100, mu=0.5)
print(aff.shape)

Increase the sparsity of the matrix by 0-clipping small values

The sparser the matrix, the higher accuracy of the GCN

# Method 1 - sparsen before row-normalization
pctile = 0.95
threshold = np.quantile(aff,pctile)
print(pctile, threshold)
aff_clip = aff.copy()
aff_clip[aff_clip < threshold] = 0
adj = sparse_tensor_from_arr(rowNormalize(aff_clip))

# Method 2 - sparsen after row-normalization
# this is the method that works for mutations
adj = rowNormalize(aff)
pctile=0.95
threshold = np.quantile(adj,pctile)
print(pctile, threshold)
adj[adj < threshold] = 0
adj = sparse_tensor_from_arr(adj)

# Train Model

import time
import argparse
import torch
import torch.nn.functional as F
import torch.optim as optim

from pygcn.utils import load_data, accuracy, balanced_accuracy
from pygcn.models import GCN
parser = argparse.ArgumentParser()
parser.add_argument('--no-cuda', action='store_true', default=False,
                    help='Disables CUDA training.')
parser.add_argument('--fastmode', action='store_true', default=False,
                    help='Validate during training pass.')
parser.add_argument('--seed', type=int, default=42, help='Random seed.')
parser.add_argument('--epochs', type=int, default=200,
                    help='Number of epochs to train.')
parser.add_argument('--lr', type=float, default=0.01,
                    help='Initial learning rate.')
parser.add_argument('--weight_decay', type=float, default=5e-4,
                    help='Weight decay (L2 loss on parameters).')
parser.add_argument('--hidden', type=int, default=16,
                    help='Number of hidden units.')
parser.add_argument('--dropout', type=float, default=0.5,
                    help='Dropout rate (1 - keep probability).')

args = parser.parse_args(['--epochs=200','--dropout=0.9','--hidden=16','--weight_decay=5e-4'])
args.cuda = not args.no_cuda and torch.cuda.is_available()
model = GCN(nfeat=features.shape[1],
            nhid=args.hidden,
            nclass=labels.max().item() + 1,
            dropout=args.dropout)
optimizer = optim.Adam(model.parameters(),
                       lr=args.lr, weight_decay=args.weight_decay)

def balanced_accuracy(output, labels):
    # output, labels - torch tensors
    preds = output.max(1)[1].type_as(labels)
    correct = preds.eq(labels).double()
    
    # Calculate the frequency of each label
    label_counts = torch.bincount(labels)
    weights = 1.0 / label_counts.float()
    
    # Apply weights to the correct predictions
    weighted_correct = correct * weights[labels]
    weighted_accuracy = weighted_correct.sum() / weights[labels].sum()
    
    return weighted_accuracy

def train(epoch):
    t = time.time()
    model.train()
    optimizer.zero_grad()
    output = model(features, adj)
    loss_train = F.nll_loss(output[idx_train], labels[idx_train])
    acc_train = balanced_accuracy(output[idx_train], labels[idx_train])
    loss_train.backward()
    optimizer.step()

    if not args.fastmode:
        # Evaluate validation set performance separately,
        # deactivates dropout during validation run.
        model.eval()
        output = model(features, adj)

    loss_val = F.nll_loss(output[idx_val], labels[idx_val])
    acc_val = balanced_accuracy(output[idx_val], labels[idx_val])
    print('Epoch: {:04d}'.format(epoch+1),
          'loss_train: {:.4f}'.format(loss_train.item()),
          'acc_train: {:.4f}'.format(acc_train.item()),
          'loss_val: {:.4f}'.format(loss_val.item()),
          'acc_val: {:.4f}'.format(acc_val.item()),
          'time: {:.4f}s'.format(time.time() - t))


def test():
    model.eval()
    output = model(features, adj)
    loss_test = F.nll_loss(output[idx_test], labels[idx_test])
    acc_test = balanced_accuracy(output[idx_test], labels[idx_test])
    acc_test_unbal = accuracy(output[idx_test], labels[idx_test])
    print("Test set results:",
          "loss= {:.4f}".format(loss_test.item()),
          "accuracy= {:.4f}".format(acc_test.item()),
          "unweighted accuracy= {:.4f}".format(acc_test_unbal.item()))

# Train model
t_total = time.time()
for epoch in range(args.epochs):
    train(epoch)
print("Optimization Finished!")
print("Total time elapsed: {:.4f}s".format(time.time() - t_total))

# Testing
test()

# FFNN control - only consider self values
adj = torch.eye(adj.size(0)).to_sparse()

# Method 2 - sparsen after row-normalization
# this is the method that works for mutations
adj = rowNormalize(aff)
pctile=0.
threshold = np.quantile(adj,pctile)
print(pctile, threshold)
adj[adj < threshold] = 0
adj = sparse_tensor_from_arr(adj)
